# Application/Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Flux Backend",
    "version": "1.0.0",
    "description": "A comprehensive API management system for environment files, web flows, and file uploads",
    "contact": {
      "name": "API Flux Team",
      "email": "support@apiflux.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.apiflux.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/web-flows/{id}/config": {
      "get": {
        "summary": "Get web flow configuration",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The web flow ID to get configuration for"
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "webFlowId": {
                          "type": "integer"
                        },
                        "nodes": {
                          "type": "array"
                        },
                        "edges": {
                          "type": "array"
                        },
                        "userId": {
                          "type": "integer"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Web flow or configuration not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Web flow configuration not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update web flow configuration",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The web flow ID to update configuration for"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nodes": {
                    "type": "array",
                    "description": "Array of flow nodes",
                    "items": {
                      "type": "object"
                    }
                  },
                  "edges": {
                    "type": "array",
                    "description": "Array of flow edges",
                    "items": {
                      "type": "object"
                    }
                  }
                },
                "required": [
                  "nodes",
                  "edges"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Web flow configuration updated successfully"
                    },
                    "config": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "webFlowId": {
                          "type": "integer"
                        },
                        "nodes": {
                          "type": "array"
                        },
                        "edges": {
                          "type": "array"
                        },
                        "userId": {
                          "type": "integer"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Web flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Web flow not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}